website/api/requirements.txt
argcomplete==1.9.2
awscli==1.14.58
base58==0.2.4
boto3==1.6.10
botocore==1.9.11
certifi==2018.1.18
cfn-flip==1.0.2
chardet==3.0.4
click==6.7
colorama==0.3.7
-e git+https://github.com/lampholder/concorde.git@aba68080fb39aa20beb02044ed2cca4e44ab3185#egg=concorde
docutils==0.14
durationpy==0.5
Flask==0.12.3
Flask-Cors==3.0.3
future==0.16.0
futures==3.2.0
hjson==3.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.19.0
MarkupSafe==1.0
matrix-client==0.0.6
placebo==0.8.1
pyasn1==0.4.2
python-dateutil==2.6.1
python-slugify==1.2.4
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
slackclient==1.1.3
toml==0.9.4
tqdm==4.19.1
troposphere==2.2.1
Unidecode==1.0.22
urllib3==1.22
websocket-client==0.47.0
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.45.1
